Remove core__study_period, additional templates
This moves the logic related to core__study_period (which was
covid specific) out of the core study.

This also makes a few additions to the generic templates, in
support of the vocab study - adding a new generic table join,
allowing generic views to specify distinct, and a new union type.